refactor: merge project General and Settings pages, upgrade deps
