+++ console-frontend/hotreload/Dockerfile
@@ -1,4 +1,9 @@
-FROM oven/bun:1-alpine
+# Use a real Node.js (>= v24.15) for `ng serve`; Angular CLI 22 rejects Bun's
+# Node-compat version (24.3.0). Bun is still used as the package manager.
+FROM oven/bun:1.3.14-alpine AS bun
+
+FROM node:24.17-alpine
+COPY --from=bun /usr/local/bin/bun /usr/local/bin/bun
 WORKDIR /app
 
 # Copy package files
