+++ console-frontend/package.json
@@ -28,46 +28,46 @@
     ]
   },
   "overrides": {
-    "@bufbuild/protobuf": "2.12.0"
+    "@bufbuild/protobuf": "2.12.1"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "21.2.15",
-    "@angular/compiler": "21.2.15",
-    "@angular/core": "21.2.15",
-    "@angular/forms": "21.2.15",
-    "@angular/platform-browser": "21.2.15",
-    "@angular/router": "21.2.15",
-    "@bufbuild/protobuf": "2.12.0",
-    "@connectrpc/connect": "2.1.1",
-    "@connectrpc/connect-web": "2.1.1",
+    "@angular/common": "22.0.2",
+    "@angular/compiler": "22.0.2",
+    "@angular/core": "22.0.2",
+    "@angular/forms": "22.0.2",
+    "@angular/platform-browser": "22.0.2",
+    "@angular/router": "22.0.2",
+    "@bufbuild/protobuf": "2.12.1",
+    "@connectrpc/connect": "2.1.2",
+    "@connectrpc/connect-web": "2.1.2",
     "@hey-api/client-fetch": "0.13.1",
-    "@nldd/design-system": "0.8.53",
+    "@nldd/design-system": "0.8.63",
     "chart.js": "4.5.1",
     "chartjs-plugin-zoom": "^2.2.0",
     "rxjs": "7.8.2",
     "tslib": "2.8.1"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "21.4.0",
-    "@angular/build": "21.2.13",
-    "@angular/cli": "21.2.13",
-    "@angular/compiler-cli": "21.2.15",
-    "@bufbuild/buf": "1.70.0",
-    "@bufbuild/protoc-gen-es": "2.12.0",
+    "@angular-eslint/builder": "22.0.0",
+    "@angular/build": "22.0.4",
+    "@angular/cli": "22.0.4",
+    "@angular/compiler-cli": "22.0.2",
+    "@bufbuild/buf": "1.71.0",
+    "@bufbuild/protoc-gen-es": "2.12.1",
     "@connectrpc/protoc-gen-connect-es": "1.7.0",
-    "@tailwindcss/postcss": "4.3.0",
-    "angular-eslint": "21.4.0",
-    "eslint": "10.4.1",
+    "@tailwindcss/postcss": "4.3.1",
+    "angular-eslint": "22.0.0",
+    "eslint": "10.5.0",
     "eslint-config-airbnb-extended": "3.1.0",
     "eslint-config-prettier": "10.1.8",
     "eslint-plugin-prettier": "5.5.6",
     "postcss": "8.5.15",
-    "prettier": "3.8.3",
+    "prettier": "3.8.4",
     "prettier-plugin-tailwindcss": "0.8.0",
-    "tailwindcss": "4.3.0",
+    "tailwindcss": "4.3.1",
     "typescript": "6.0.3",
-    "typescript-eslint": "8.60.0",
-    "vitest": "4.1.7"
+    "typescript-eslint": "8.62.0",
+    "vitest": "4.1.9"
   }
 }
