--- console-frontend/src/app/project-settings/project-settings.component.html
@@ -1,156 +1,0 @@
-<div class="card">
-  <!-- Header -->
-  <div class="card-header">
-    <h1 class="text-3xl font-semibold">Project</h1>
-    <p class="mt-1 text-gray-600 dark:text-gray-300">Manage your project's information</p>
-  </div>
-
-  <div class="p-6">
-    <!-- Error Message -->
-    @if (error()) {
-      <div class="bg-danger-50 dark:bg-danger-950 mb-4 rounded-md p-4">
-        <p class="text-danger-800 dark:text-danger-200">{{ error() }}</p>
-      </div>
-    }
-
-    <!-- Loading State -->
-    @if (loading()) {
-      <div class="py-8 text-center">
-        <p class="text-gray-600 dark:text-gray-300">Loading project...</p>
-      </div>
-    }
-
-    <!-- Project Information -->
-    @if (project()) {
-      <div class="space-y-6">
-        <!-- Project ID -->
-        <div>
-          <div class="mb-2 block font-medium text-gray-700 dark:text-gray-300">Project ID</div>
-          <p
-            class="rounded-md bg-gray-50 px-4 py-2 font-mono text-sm text-gray-900 dark:bg-gray-900 dark:text-white"
-          >
-            {{ project()!.id }}
-          </p>
-        </div>
-
-        <!-- Project Name (immutable) -->
-        <div>
-          <div class="mb-2 block font-medium text-gray-700 dark:text-gray-300">
-            Name <span class="font-normal text-gray-500">(cannot be changed)</span>
-          </div>
-          <p
-            class="rounded-md bg-gray-50 px-4 py-2 font-mono text-sm text-gray-900 dark:bg-gray-900 dark:text-white"
-          >
-            {{ project()!.name }}
-          </p>
-        </div>
-
-        <!-- Project Alias (editable) -->
-        <div>
-          <div class="mb-2 block font-medium text-gray-700 dark:text-gray-300">
-            Alias <span class="font-normal text-gray-500">(name that is displayed)</span>
-          </div>
-          @if (isEditing()) {
-            <div class="flex gap-2">
-              <nldd-text-field
-                appAutofocus
-                [value]="editingAlias()"
-                (input)="editingAlias.set($any($event.target).value)"
-                class="flex-1"
-                (keyup.enter)="saveEdit()"
-                (keyup.escape)="cancelEdit()"
-              ></nldd-text-field>
-              <nldd-button
-                type="button"
-                (click)="saveEdit()"
-                [disabled]="loading()"
-                text="Save"
-                start-icon="check-mark"
-                variant="primary"
-              ></nldd-button>
-              <nldd-button
-                type="button"
-                (click)="cancelEdit()"
-                [disabled]="loading()"
-                text="Cancel"
-                start-icon="dismiss-small"
-                variant="secondary"
-              ></nldd-button>
-            </div>
-          } @else {
-            <div class="flex items-center gap-3">
-              <p>{{ project()!.alias }}</p>
-              <nldd-button
-                type="button"
-                (click)="startEdit()"
-                [disabled]="loading()"
-                text="Rename"
-                start-icon="pencil"
-                variant="secondary"
-                size="sm"
-              ></nldd-button>
-            </div>
-          }
-        </div>
-
-        <!-- Created Date -->
-        <div>
-          <div class="mb-2 block font-medium text-gray-700 dark:text-gray-300">Created</div>
-          <p>
-            {{ formatDate(project()!.created) }}
-          </p>
-        </div>
-
-        <!-- Danger Zone -->
-        <div
-          class="bg-danger-50 dark:bg-danger-900/20 col-span-full rounded-md border border-rose-200 p-4 dark:border-rose-900"
-        >
-          <h3 class="text-danger-800 dark:text-danger-200 mb-2 text-lg font-medium">Danger zone</h3>
-          <p class="text-danger-700 dark:text-danger-300 mb-4">
-            Deleting a project is permanent and cannot be undone.
-          </p>
-          <nldd-button
-            (click)="showDeleteModal.set(true)"
-            text="Delete this project"
-            variant="destructive"
-          ></nldd-button>
-        </div>
-      </div>
-    }
-  </div>
-</div>
-
-<!-- Delete Project Modal -->
-<nldd-modal-dialog
-  #deleteDialog
-  appDialogSync
-  [show]="showDeleteModal()"
-  text="Delete project"
-  variant="alert"
-  (open)="onDeleteModalOpen()"
-  (close)="showDeleteModal.set(false)"
->
-  <nldd-spacer size="24"></nldd-spacer>
-
-  <p>
-    Are you sure you want to delete project
-    <strong class="font-semibold">{{ project()?.alias || project()?.name }}</strong
-    >?
-  </p>
-  <p class="text-danger-600 dark:text-danger-400 mt-2">This cannot be undone.</p>
-
-  <nldd-button
-    slot="actions"
-    type="button"
-    (click)="deleteProject()"
-    text="Yes, delete this project"
-    variant="destructive"
-  ></nldd-button>
-  <nldd-button
-    slot="actions"
-    type="button"
-    (click)="showDeleteModal.set(false)"
-    text="Cancel"
-    variant="secondary"
-  ></nldd-button>
-</nldd-modal-dialog>
