--- console-frontend/src/app/project-settings/project-settings.component.ts
@@ -1,183 +1,0 @@
-import {
-  Component,
-  inject,
-  signal,
-  OnInit,
-  ChangeDetectionStrategy,
-  CUSTOM_ELEMENTS_SCHEMA,
-  viewChild,
-  ElementRef,
-} from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { ConnectError, Code } from '@connectrpc/connect';
-import { create } from '@bufbuild/protobuf';
-import { firstValueFrom } from 'rxjs';
-import AutofocusDirective from '../autofocus.directive';
-import { TitleService } from '../title.service';
-import { ToastService } from '../toast.service';
-import { OrganizationDataService } from '../organization-data.service';
-import DialogSyncDirective from '../dialog-sync.directive';
-import focusFirstModalInput from '../modal-focus';
-import { PROJECT } from '../../connect/tokens';
-import {
-  GetProjectRequestSchema,
-  UpdateProjectRequestSchema,
-  DeleteProjectRequestSchema,
-  type Project,
-} from '../../generated/v1/project_pb';
-import { formatDate as formatDateUtil } from '../utils/date-format';
-
-@Component({
-  selector: 'app-project-settings',
-  imports: [DialogSyncDirective, AutofocusDirective],
-  schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  templateUrl: './project-settings.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export default class ProjectSettingsComponent implements OnInit {
-  private titleService = inject(TitleService);
-
-  private route = inject(ActivatedRoute);
-
-  private router = inject(Router);
-
-  private projectClient = inject(PROJECT);
-
-  private toastService = inject(ToastService);
-
-  private organizationDataService = inject(OrganizationDataService);
-
-  projectId = signal<string>('');
-
-  project = signal<Project | undefined>(undefined);
-
-  isEditing = signal(false);
-
-  editingAlias = signal('');
-
-  loading = signal(false);
-
-  error = signal<string | null>(null);
-
-  showDeleteModal = signal<boolean>(false);
-
-  constructor() {
-    this.titleService.setTitle('Project Settings');
-  }
-
-  async ngOnInit() {
-    const id = this.route.snapshot.params['id'];
-    if (id) {
-      this.projectId.set(id);
-      await this.loadProject(id);
-    }
-  }
-
-  private async loadProject(projectId: string) {
-    this.loading.set(true);
-    this.error.set(null);
-    try {
-      const request = create(GetProjectRequestSchema, { projectId });
-      const response = await firstValueFrom(this.projectClient.getProject(request));
-      if (response.project) {
-        this.project.set(response.project);
-      }
-    } catch (err) {
-      this.error.set(
-        err instanceof Error ? `Failed to load project: ${err.message}` : 'Failed to load project',
-      );
-    } finally {
-      this.loading.set(false);
-    }
-  }
-
-  startEdit() {
-    const currentProject = this.project();
-    if (currentProject) {
-      this.isEditing.set(true);
-      this.editingAlias.set(currentProject.alias || currentProject.name);
-    }
-  }
-
-  cancelEdit() {
-    this.isEditing.set(false);
-    this.editingAlias.set('');
-  }
-
-  async saveEdit() {
-    const currentProject = this.project();
-    const aliasToSave = this.editingAlias();
-
-    if (!aliasToSave.trim() || !currentProject) {
-      return;
-    }
-
-    this.loading.set(true);
-    this.error.set(null);
-
-    try {
-      const request = create(UpdateProjectRequestSchema, {
-        projectId: currentProject.id,
-        alias: aliasToSave.trim(),
-      });
-      await firstValueFrom(this.projectClient.updateProject(request));
-
-      this.project.set({
-        ...currentProject,
-        alias: aliasToSave.trim(),
-      });
-      this.organizationDataService.updateProjectAlias(currentProject.id, aliasToSave.trim());
-      this.isEditing.set(false);
-      this.editingAlias.set('');
-    } catch (err) {
-      this.error.set(
-        err instanceof Error
-          ? `Failed to update project alias: ${err.message}`
-          : 'Failed to update project alias',
-      );
-    } finally {
-      this.loading.set(false);
-    }
-  }
-
-  async deleteProject() {
-    const currentProject = this.project();
-    if (!currentProject) return;
-
-    try {
-      const request = create(DeleteProjectRequestSchema, {
-        projectId: currentProject.id,
-      });
-
-      await firstValueFrom(this.projectClient.deleteProject(request));
-
-      this.showDeleteModal.set(false);
-      this.toastService.success(`Project '${currentProject.name}' deleted`);
-
-      // Reload project data to update the selector modal
-      await this.organizationDataService.reloadProjectsAndNamespaces();
-
-      this.router.navigate(['/projects']);
-    } catch (err) {
-      this.showDeleteModal.set(false);
-      if (err instanceof ConnectError && err.code === Code.FailedPrecondition) {
-        this.error.set('Delete all namespaces in this project before deleting the project.');
-      } else {
-        this.error.set(
-          err instanceof Error
-            ? `Failed to delete project: ${err.message}`
-            : 'Failed to delete project',
-        );
-      }
-    }
-  }
-
-  readonly formatDate = formatDateUtil;
-
-  deleteDialogRef = viewChild<ElementRef<HTMLElement>>('deleteDialog');
-
-  onDeleteModalOpen(): void {
-    const el = this.deleteDialogRef()?.nativeElement;
-    if (el) focusFirstModalInput(el);
-  }
-}
